Swift: move logging to a common directory
