+++ swift/extractor/trap/BUILD.bazel
@@ -23,7 +23,10 @@ genrule(
         "--schema=$(location //swift:schema)",
         "--script-name=codegen/codegen.py",
     ]),
-    exec_tools = ["//misc/codegen", "//swift:schema"],
+    exec_tools = [
+        "//misc/codegen",
+        "//swift:schema",
+    ],
 )
 
 filegroup(
@@ -49,7 +52,7 @@ swift_cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//swift/extractor/infra/file",
-        "//swift/extractor/infra/log",
+        "//swift/log",
         "@absl//absl/numeric:bits",
     ],
 )
