+++ swift/log/SwiftLogging.h
